requirements.txt
python-dotenv>=1.0.0
requests==2.32.3
volcengine==1.0.188
volcengine-python-sdk==4.0.34
pydantic>=2.0.0
tenacity>=8.0.0
openai>=1.0.0
